replace more unit tests so they use numba util
