+++ autoarray/inversion/pixelization/mappers/mapper_numba_util.py
@@ -1,8 +1,12 @@
 import numpy as np
 from typing import Tuple
 
+from autoconf import conf
+
 from autoarray import numba_util
 from autoarray.inversion.pixelization.mesh import mesh_numba_util
+
+from autoarray import exc
 
 
 @numba_util.jit()
@@ -245,3 +249,105 @@ def remove_bad_entries_voronoi_nn(
         pix_weights_for_sub_slim_index[ind][0] = 1.0
 
     return pix_weights_for_sub_slim_index, pix_indexes_for_sub_slim_index
+
+
+def pix_size_weights_voronoi_nn_from(
+    grid: np.ndarray, mesh_grid: np.ndarray
+) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
+    """
+    Returns the mappings between a set of slimmed sub-grid pixels and pixelization pixels, using information on
+    how the pixels hosting each sub-pixel map to their closest pixelization pixel on the slim grid in the data-plane
+    and the pixelization's pixel centres.
+
+    To determine the complete set of slim sub-pixel to pixelization pixel mappings, we must pair every sub-pixel to
+    its nearest pixel. Using a full nearest neighbor search to do this is slow, thus the pixel neighbors (derived via
+    the Voronoi grid) are used to localize each nearest neighbor search by using a graph search.
+
+    Parameters
+    ----------
+    grid
+        The grid of (y,x) scaled coordinates at the centre of every unmasked pixel, which has been traced to
+        to an irgrid via lens.
+    slim_index_for_sub_slim_index
+        The mappings between the data slimmed sub-pixels and their regular pixels.
+    mesh_grid
+        The (y,x) centre of every Voronoi pixel in arc-seconds.
+    neighbors
+        An array of length (voronoi_pixels) which provides the index of all neighbors of every pixel in
+        the Voronoi grid (entries of -1 correspond to no neighbor).
+    neighbors_sizes
+        An array of length (voronoi_pixels) which gives the number of neighbors of every pixel in the
+        Voronoi grid.
+    """
+
+    try:
+        from autoarray.util.nn import nn_py
+    except ImportError as e:
+        raise ImportError(
+            "In order to use the Voronoi pixelization you must install the "
+            "Natural Neighbor Interpolation c package.\n\n"
+            ""
+            "See: https://github.com/Jammy2211/PyAutoArray/tree/main/autoarray/util/nn"
+        ) from e
+
+    max_nneighbours = conf.instance["general"]["pixelization"][
+        "voronoi_nn_max_interpolation_neighbors"
+    ]
+
+    (
+        pix_weights_for_sub_slim_index,
+        pix_indexes_for_sub_slim_index,
+    ) = nn_py.natural_interpolation_weights(
+        x_in=mesh_grid[:, 1],
+        y_in=mesh_grid[:, 0],
+        x_target=grid[:, 1],
+        y_target=grid[:, 0],
+        max_nneighbours=max_nneighbours,
+    )
+
+    bad_indexes = np.argwhere(np.sum(pix_weights_for_sub_slim_index < 0.0, axis=1) > 0)
+
+    (
+        pix_weights_for_sub_slim_index,
+        pix_indexes_for_sub_slim_index,
+    ) = remove_bad_entries_voronoi_nn(
+        bad_indexes=bad_indexes,
+        pix_weights_for_sub_slim_index=pix_weights_for_sub_slim_index,
+        pix_indexes_for_sub_slim_index=pix_indexes_for_sub_slim_index,
+        grid=np.array(grid),
+        mesh_grid=np.array(mesh_grid),
+    )
+
+    bad_indexes = np.argwhere(pix_indexes_for_sub_slim_index[:, 0] == -1)
+
+    (
+        pix_weights_for_sub_slim_index,
+        pix_indexes_for_sub_slim_index,
+    ) = remove_bad_entries_voronoi_nn(
+        bad_indexes=bad_indexes,
+        pix_weights_for_sub_slim_index=pix_weights_for_sub_slim_index,
+        pix_indexes_for_sub_slim_index=pix_indexes_for_sub_slim_index,
+        grid=np.array(grid),
+        mesh_grid=np.array(mesh_grid),
+    )
+
+    pix_indexes_for_sub_slim_index_sizes = np.sum(
+        pix_indexes_for_sub_slim_index != -1, axis=1
+    )
+
+    if np.max(pix_indexes_for_sub_slim_index_sizes) > max_nneighbours:
+        raise exc.MeshException(
+            f"""
+            The number of Voronoi natural neighbours interpolations in one or more pixelization pixel's 
+            exceeds the maximum allowed: max_nneighbors = {max_nneighbours}.
+
+            To fix this, increase the value of `voronoi_nn_max_interpolation_neighbors` in the [pixelization]
+            section of the `general.ini` config file.
+            """
+        )
+
+    return (
+        pix_indexes_for_sub_slim_index,
+        pix_indexes_for_sub_slim_index_sizes,
+        pix_weights_for_sub_slim_index,
+    )
