+++ test_autoarray/inversion/pixelization/mappers/test_delaunay.py
@@ -28,7 +28,7 @@ def test__pix_indexes_for_sub_slim_index__matches_util(grid_2d_sub_1_7x7):
     (
         pix_indexes_for_sub_slim_index_util,
         sizes,
-    ) = aa.util.mapper.pix_indexes_for_sub_slim_index_delaunay_from(
+    ) = aa.util.mapper_numba.pix_indexes_for_sub_slim_index_delaunay_from(
         source_plane_data_grid=np.array(mapper.source_plane_data_grid),
         simplex_index_for_sub_slim_index=simplex_index_for_sub_slim_index,
         pix_indexes_for_simplex_index=pix_indexes_for_simplex_index,
